feat: :sparkles: modified J2Reader so it supports file.yml.j2 and file.json.j2
Added abstract method read_content() to ConfFileReader. This method takes a content string and reads the content as a configuration according to the subclass. For J2Reader, this method will also make use of a mapping dictionary that store key values to be rendered along side the content string. In the read_file() method of J2Reader, will call a read_content method from the subclass appropriate to the file extension.